Add optimization option to reverse thread group (#1986)
This commit add optimization option to reverse thread group.

When the option is set, LGC will generate code to reverse thread group
order by a driver given control value. Driver is responsible to fill the
internal buffer with either 0 or 1, and LGC will load that value to
determine whether gl_WorkGroupId needs to be reversed.

It is expected that performance can improve for some compute shaders
that will access the same resource between each dispatch. By reversing
the order, the thread groups will begin accessing resource data that is
still in cache due to the earlier dispatch.