Add mesh shader support: Fix an issue of primitiveCount=0
There is a test case. The vertex count emitted by mesh shader is not 0
while the primitive count is 0. Thus, HW will malfunction after the case
finishes. We have to detect if any of vertex and primitive count is 0
and make a guard to force both of them to be 0 in such case.