Add mesh shader support: Fix an issue of unlinked mesh/fragment pipeline
For unlinked mesh/fragment pipeline, we don't go into right code path to
fetch the mapped location for a per-primitive input of fragment shader.
The original condition was to check fragment shader's previous stage. But in
unlinked pipeline, the previous stage is invalid stage, we have to check
unlinked state.