Return error message when resource is not found
When creating pipeline, there is no matching resource we still return a
correct result, which is not expected.

This change is fixing this behavior, but it is invalid for relocatable
shader, because we can't know if it is compiling a single shader with
amdllpc.

In addition, fixes the CacheAccessor destructor error.