lgc: streamline generic FS input IR
LGC IR now contains one of the following operations for reading an FS
input:

- @lgc.input.import.generic.%T%(i32 location, i32 elemIdx)
  Used for per-primitive inputs, i.e. in mesh shader pipelines.

- @lgc.input.import.interpolant.%Type%(i32 location, i32 locOffset,
          i32 elemIdx, i32 interpMode, <2 x float> | i32 interpValue)
  Used for not-per-primitive (per-vertex / interpolated) inputs.
  interpMode is one of:
    - InterpModeSmooth: interpolate in the normal way, using
      barycentrics from interpValue
    - InterpModeCustom: retrieve the values of one of the vertices;
      interpValue is the (hardware) vertex index (0, 1, or 2)
    - InterpModeFlat: flat shading, interpValue is ignored

This simplifies the IR relative to the status quo, where the handling of
interpolation location (center / centroid / sample) is spread across the
frontend building of IR and PatchInOutImportExport. Now, this is handled
consistently in the frontend. Note the associated slight reduction in
code size.

The alternative would have been to always handle locations in
PatchInOutImportExport and streamline the IR to allow (mode, location,
aux) tuples to be represented such that the required information
survives in the IR.

The disadvantage of the chosen solution is that it results in slightly
larger IR, since operations to read barycentrics are explicitly
included. The advantage is that the definition of the IR is simpler.

v2:
- force interpMode to "Flat" for per-primitive inputs in
  PatchInOutImportExport