Opaque pointers: Add command line flag to enable opaque-pointers
Opaque pointer flag which is defined in LLVM is by default enabled. For
transition time, we need to have new flag which will be by default
disabled. This flag will be used in LIT tests to test opaque and
non-opaque pointers. After merging all the patches this flag will
be changes to true (by default) and removed if regression does not show
up.