Make internalRtShaders a pipeline option
Currently the usage of internal RT shader intrinsics is not
captured by the pipeline state, and there is no way to activate
these from the command line. This makes it impossible to
recompile an internal RT shader, such as BuildBVH on the command
line.

Move the isInternalRtShader module option to a pipeline option.
Add a command line option -enable-internal-rt-shaders.

Load and dump internalRtShaders option to .pipe files.
An XGL change is required for the dump state to be accurate.