Relocatable shaders support compact descriptor
For shader compilation with no user data layout provided, we don't

know if the buffer is a dynamic descriptor, so we need to get the

two descriptors(DescriptorBuffer, DescriptorBufferCompact), When

linking, we determine the value.

Add a internal resource node which descriptor set is -1 and type

is DescriptorBufferCompact, we use spillTbale to load it.