Opaque Pointers: Handle name mangling.
With opaque pointers we are not able to get base type of the pointer, so
while creating mangled function name we are able only to indicate that
we are dealing with pointer type (func_name.p7 instead of
func_name.p7v4f32). This is way return type of the function has been
added while mangling the function name.