Add aggressiveInvariantLoads and disableInvariantLoads options
Add patching pass to modify invariant load metadata based on
aggressiveInvariantLoads and disableInvariantLoads in
PipelineShaderOptions.

If aggressiveInvariantLoads is set then all loads in a shader are
marked invariant if there are no memory writes in the shader.
Note: barriers, discards, and inline asm are considered writes to
memory so this generally safe.

If disableInvariantLoads is set then all invariant load metadata
is striped from loads in the shader.
This is useful for optimising edges where sink in causing unwanted
register pressure increases.