Fix an issue of s_sendmsg in mesh shader
M0 implicitly used in s_sendmsg is SGPR. Therefore, the value written to
M0 should be an uniform one. Therefore, we must promote vertex/primitive
count to SGPR. This is because LLVM backend is not smart enough to
translate IR select to s_cselect. Rather, it uses v_cndmask, which
eventually converts vertex/primitive count to VGPR. And LLVM backend
reports illegal copy.