Tag MultiDispatchInfo argument for LDS spilling
Tag the position of MultiDispatchInfo argument, so the backend knows
which sgpr needs to be preloaded for COMPUTE_PGM_RSRC2.tg_size_en
(Work-Group Info).

The WG info is needed to generate correct offsets in the spill code
if the workgroup size is larger than wavefront size, as it uses
DS_READ_ADDTID/DS_WRITE_ADDTID instructions.