lgc: start adding tests for individual passes
Begin with a simple test case for the PatchReadFirstLane pass; and yes,
this was in part motivated by the weaknesses in the pass that this test
exposes.

As noted in the documentation, this approach will not work as-is for the
more complex passes that rely on outside-IR state. But it does apply to
some of our more self-contained passes, and that's a good start.

Fix a latent copy&paste error in the test CMakeLists.txt to correctly
handle adding the test subdirectories.

Note that I'm using a slight modification to update_test_checks.py, see
https://reviews.llvm.org/D136329.