Heuristically enable aggressive invariant loads optimisation.
On GFX10+ aggressive invariant load labelling is beneficial to
VGPR pressure on many shader types with very few outliers.
(Based on an analysis of common game shaders.)

Change the aggressiveInvariantLoads shader option to a tristate
with the default value being Auto, mean heuristics are applied.
With explicit Enable and Disable values are also possible.