Fix user data limit issue of task shader
For task shader, the user data limit for task shader is the same as
compute shader because it is mapped to HW CS eventually. The check was
missing originally, leading to CTS VK.binding_model.*.task.* failures.

Also, fix an assert when checking if user data index is in range. The
previous check is unclear and is incorrect.