Add new interface CreateIntegerDotProductNew() (#2084)
Re-implement interger dot product in `CreateIntergerDotProductNew()`.
The factor inputs are `<N x iM>` of the same type, where N can be
arbitrary and M must be 4, 8, 16, 32 or 64.

Note: The old interface is still in use in this change. I'm going to
replace the old one after this PR is ported to the internal repo.