Add SRD stride variable in resource node struct
The SRD stride variable is used to load a descriptor with an offset value
rather than a fixed image descritor size. For example, if the client uses
full combined image sampler descriptor set layout with multi-planar
YCbCr image, the descriptor stride between each planes should be the size of
image descriptor and sampler descriptor.

The load with offset changes would be in the later commits.