lgc: support the 'lgc-pipeline-state' analysis in -passes
We have some function passes, like PatchBufferOp, that require
PipelineState, which is obtained via a module analysis. The analysis
proxy cannot run a module analysis from a function pass, so any
standalone tests of such passes must explicitly ensure that the
PipelineState analysis is run before. The canonical way to achieve this
is by adding 'require<lgc-pipeline-state>' to the pass pipeline, as is
shown in the very simple new test.