Add !lgc.xfb.state metadata
We're now preserving XFB state in explicit metadata so that we can
handle the case where there is an XFB output that is never written to.
Meanwhile, the the redudant XFB-related information of
`lgc::ResourceUsage::{enableXfb,xfbStrides,streamXfbBuffers}` is
removed.
The detailed dicussions are in issue(https://github.com/GPUOpen-Drivers/llpc/issues/2109).

See:dEQP-VK.transform_feedback.simple.multiquery_omit_write*
    dEQP-VK.transform_feedback.simple.query_omit_write_0*