Remove use of legacy passes being removed upstream (#2136)
Upstream llvm is removing some of the legacy passes.
LLPC has moved over to the new pass manager, and we should remove the code for
the legacy pass manager. This is quite a big piece of work.
In the meantime, remove AggressiveInstCombiner as it is disappearing from
upstream llvm, in preparation for subsequent merge.