Fix an issue of unspecified outputs that will be exported to FS
We encounter a demo. In this demo, an output is not written by VS while
its corresponding input is read by FS. The value then involves in some
operations in FS and lead to corruption. Of course, this is surely an
App issue violating GLSL spec requirements. But we can do something to
accept this case.