Add tool/update_llpc_test_checks.py
The goal is to allow us to test the frontend (LLPC) IR emission with
tests that are more meaningful and also easier to maintain and update
than the status quo.

This is obviously built on LLVM's update_test_checks tools, so a
reasonable question is whether we can re-use those tools. The existing
UTC tools themselves make too many assumptions about how the underlying
tool is invoked and about the input language(s), so a new tool is
appropriate.

Since the utils Python scripts aren't typically seen as a publically
exposed interface of LLVM, I decided to just copy the helper library.
This makes it easier for us to make the changes in the library that we
need (and there are some) and should avoid random breakage due to
upstream changes.

I've changed an arbitrary existing test to demonstrate the kind of test
file that results.

v2:
- fix typos