Applying VgprLimit to all shaders
If a --vgpr-limit is passed it will be applied to all shaders even if they
have their own VGPR limit set. The lower of the two limits will be in
effect in such a situation.

This avoids situations where some shaders used more VGPRs than allowed and
makes this options' semantics less surprising.