[CI] Rename remotes to origin
Calling the remote 'origin' seems to fix issues when pulling with
submodules. I couldn’t reproduce this outside of docker, with my newer
git version, so probably this was a bug that is now fixed in git.

I tested with Ubuntu 22.04 as the base image, so the issue is still
there, even if we update the base image.