lgc: set cumode target feature based on the wgpMode shader option
Correctly setting "-cumode" is necessary for correctness when WGP mode
is used because it affects the memory model legalizer (i.e., WGP mode
requires more cache invalidations).

It seems we got lucky so far, probably in part because we don't use WGP
mode that much. I only noticed this by chance when looking through the
code.