SPIRVReader: cleanup OpAccessChain translation
The old approach of remembering "splits" in the access chain was fragile
and difficult to extend.

The new approach is based on the simple observation that each step of
the OpAccessChain can in principle be translated on its own. We do
collect gepIndices across multiple steps as an obvious optimization, but
whenever the chain needs to be split, we do so immediately.