Relocatable shaders support compact descriptor (#2106)
* Relocatable shaders support compact descriptor

For shader compilation with no user data layout provided, we don't

know if the buffer is a dynamic descriptor, so we need to get the

two descriptors(DescriptorBuffer, DescriptorBufferCompact), When

linking, we determine the value.

Add a internal resource node which descriptor set is -1 and type

is DescriptorBufferCompact, we use spillTbale to load it.

* Support uber fetch shader

In relocatable shader mode, Current fetch shader exists an issue

which generate too many function arguments when input attributes

are too many, which is leading some failures.

Enabling uber shader, driver will write all input descriptors into

user data, we read these resources to load vertex.

* Use an s_load_b128 instruction to load the UberFetchShaderAttribInfo

What's stored in the user data is a plain 64-bit pointer, use

s_load_b128 to replace s_buffer_load_b128.

* Fix the instance divisor

When instance is big enough, we may get an incorrect result. Change the
relevant code on driver side to ensure that the instance divisor is not
passed in.

Add an assertion to ensure uber fetch shader only supports the formats
after gfx9.