Remove old pass-manager code
Remove the wrappers for the legacy pass manager.
I needed to change one test. -stop-after didn’t work anymore, because
the legacy, code-gen pass manager didn’t find a pass with that name
(because it only exists as a npm pass now).
I think that could be improved by not running code-gen passes when a
-stop-after option is given.

One thing that’s missing in this change to get rid of the legacy code is
moving the code from runImpl functions into the run functions.