lgc: Prepare VertexFetch for Builder::setShaderStage removal
VertexFetch wants to create a load of a buffer descriptor, so it creates
a BuilderImpl and then uses its Builder virtual method. That is a bit
hairy, for the following reason:

We can think of all the lgc.create.* calls that BuilderRecorder
generates as an "LGC dialect", albeit not using Nicolai's new dialect
infrastructure. There is currently a single pass (BuilderReplayer) to
lower ops in that dialect to something else. Then, because the
problematic bit of code (pre-gfx9 über fetch shader support) wants to
use the "load buffer desc" dialect op, it would have to go before the
pass that lowers it. Using a new instance of BuilderImpl here is not
supposed to work, although I think we get away with it in this
particular case.

The immediate problem I need to fix here is that this VertexFetch code
calls Builder::setShaderStage(), and I am about to remove it as we no
longer need the front-end to call it. So this is a shorter term fix to
cast to BuilderImplBase and call setShaderStage on that.  To get that to
work, I had to move BuilderImpl.h into an include directory so it is
visible outside the builder directory.

Longer term we would want to be using a "load buffer descriptor" dialect
op, and ensure we have the correct ordering of dialect lowering.