Add float atomic min/max intrinsic support for GPURT
This commit adds support for several GPURT intrinsic that perform float
atomic min/max operations on buffer and LDS.

Known issue: For LDS intrinsic (AmdExtAtomicLdsFMin/AmdExtAtomicLdsFMax),
we need to pass in the `groupshared` variable as function parameter,
however, HLSL does not allow function parameter to be `groupshared`, but
can call a function with a groupshared argument. SPIRVReader will always
assert because of the mismatch function call. We fix the mismatch by
mutating the parameter in later pass.