NGG coding rework: Rename vertexId to vertexIndex
The reason is that vertex ID is confusing when there is HW global vertex
ID passed by GE. The vertex ID in NGG is actually a relative index
of a vertex in a NGG subgroup. Also, mesh shader uses the term "indices"
to specify a primitive topology. We can borrow this idea.