lgc/PatchBufferOp: explicitly remove vector-of-pointer support
Vector-of-pointer support had a number of problems:

- it was incomplete: at least shufflevector support was missing
- it didn't support the case of non-uniform descriptors within a single
  vector of pointers

I ran into this when trying to remove the scalarizer pass from the patch
pipeline.

These issues could in theory be fixed. However, vectors-of-pointers are
somewhat surprising, aren't supported by SPIR-V, and are generally not
well tested and supported (e.g., some combines may not be firing for
them, alias analysis may get confused, etc.).

I think it is best we make the conscious decision to just not use or
support them at all.