Opaque Pointers: Try to add missing zero-indexs between GEP and Global Value
While doing translation to LLVM-IR for OpStore or OpLoad which
are doing store/load of the whole structure:

example with OpLoad
%_struct_57 = OpTypeStruct %_arr_long_uint_16_0 %_arr_long_uint_32_0
OpLoad %_struct_57 %62

LLPC is creating multiple store/load instructions to handle each single element
of the structure. This is done by recursively unpacking structure type, until
basic type is not found (i32, i64, float, ...). While unpacking each level of
the structure, the GEP instruction is created. In case of nested strutures LLPC
is trying to create zero-index GEP which will not be created for opaque pointers,
since it is not changing the pointer address. GEP instruction will just fold to
Global Variable instruction.

Later while doing Lowering of Global Variables, LLPC is using these indices from
GEP to unpack Global Variable type and it's metadata. If types of GV and GEP are
different (which means that some of the zero-index elements where removed) then
we are accessing wrong places of Global Value's type and metadata.

This change is creating new GEP instruction with _correct_ number of indices so
that types of GEP and GV are the same.