Make a test more robust
This test fails based on the LLVM version because different versions of
LLVM schedule instructions differently which throws off the matching.
Since the point of that sub-test is to make sure the relocated
descriptor stride is substituted correctly, and the resulting code should
have exactly one s_mul instruction, we can simplify the test to make it
more robust.