Enable uber fetch shader for gfx8 and gfx9
In the earliest version, we added an assert for gfx9 and older, which

causes a test failure in internal pre-checkin. Therefore it was diabled.

However, a lot tests that require uber fetch shader are being added on

gfx8, so we have to enable it. For the unsupported formats (10_10_10_2),

we rely on the driver to fallback to not using the uber fetch shader

when these formats are used.