Fix an issue of GFX11 GS transform feedback
There are tests that do rasterization yet still do transform feedback.
The numbers of primitives generated are queried both for raster stream
and for non-raster stream. The GLSL source is something like this:

  layout(points) in;
  layout(points, max_vertices = 2) out;
  layout(xfb_buffer = 0, xfb_offset = 0, xfb_stride = 16,
         location = 0, stream = 1) out vec4 xfb;
  void main (void) {
    gl_PointSize = 1.0;
    xfb = vec4(42);
    EmitStreamVertex(0);
    EndStreamPrimitive(0);
    EmitStreamVertex(1);
    EndStreamPrimitive(1);
  }

There are two streams, one is for rasterization but is not associated
with XFB and the other is dedicated for XFB. The number of generated
primitives are finally queried. Therefore, we must treat vertex streams
as active if they are associated with XFB buffer or is raster stream
itself. Otherwise, the raster stream is not processed.