Update test to avoid cat and pipes
Using cat as part of a pipe (foo | cat | FileCheck) is problematic for any setup
that uses the LIT_USE_INTERNAL_SHELL=1 environment variable.
This is commonly used for some not-very POSIX compliant platforms.

Update the test to go via an intermediate file - this still achieves the
original intent of the cat (to stop update from changing some check lines).