NGG coding rework: Don't conditionally create blocks
This change is to create all possible blocks unconditionally and make
some of them empty if the processing should be skipped. The empty blocks
will be removed by SimplifyCFG. The finally generated codes are the
same.

The reason for this is the conditonal block creation makes the processing
duplicated and unclear. This is very error-prone when we have multiple
conditions that controlling block creation and their branching
relationship.

The function buildPassthroughPrimShader is modified in this change. Will
continue to change buildPrimShader and buildPrimShaderWithGs in separate
changes.

The change also adds a helper distributePrimitiveId. This function will
handle all processing regarding primitive ID distribution.