NggPrimShader: Improve attribute-through-memory code generation
Previously, we would export attributes on gfx11 like:

    ; v3 is threadIdInSubgroup
    s_add_i32 s8, s9, 0x200
    buffer_store_b128 v[10:13], v3, s[0:3], s8 idxen glc slc
    buffer_store_b128 v[32:35], v3, s[0:3], s9 idxen glc slc

With this change, we export attributes like:

    buffer_store_b128 v[10:13], v3, s[0:3], s9 idxen offset:16 glc slc
    buffer_store_b128 v[32:35], v3, s[0:3], s9 idxen glc slc

That is, we lean on the buffer swizzling to avoid setting up different
offsets for the different locations.

The change also cuts a few SALU instructions from the buffer descriptor
fixup.

I started this hoping to get rid of the index VGPR, but have run into
issues that may or may not be a hardware problem. In the meantime, this
is a smaller improvement that should just work.