builder: Fix wave64 subgroup shuffle
Fix logic for subgroup shuffle.

Since a `ds_bpermute` takes the source
index and performs a `% 32`, it only sees
the lower lanes as source. In order
to get the values from the higher lanes,
we perform a `v_permlane64`. Now,
the values from the upper lanes are
accessible to the lower ones.

We perform two `ds_bpermute` instructions and then select
the correct results for all the lanes.

Initially, I thought the results are the
bpermuted values from the lower and
the higher lanes, respectively.

However, they are the values from the same half and the other half. So
instead of checking if the source is
in the upper half, we now check if the source is in the same half.
If it is, we can select the values from the first bpermute.
Otherwise we take the second bpermute result.