Support non-zero index component outputs exported to XFB
The component index of an output is also allowed to be non-zero when
exporting to a transform feedback buffer. But we assume the index is
zero-based. This case triggers the assertion
`assert(outputLocInfoMap.count(origLocInfo));` in the output packing
path in `PatchCopyShader::exportOutput`. The reason is that the keys of
the two output maps are not equal any more. `outputLocInfoMap` is build
from the generic output calls that include the info of the component
index, while `locInfoXfbOutInfoMap` is built from xfb output calls that
lack the info of component index. In order to get component index for
`locInfoXfbOutInfoMap`, we need modify the interface
`CreateWriteXfbOutput` to pass the component index through function
argmument.
Update PipelineGsTess_TestInOutPacking.pipe to cover this case.