Fix setting subgroupSize
setShaderDefaultWaveSize set the wave size but missed to also set the
subgroup size in some cases.
When a shader on gfx10.3+ did not observe the subgroup size and the
workgroup size was >64, a shader ended up with waveSize=64 and
subgroupSize=32.

This patch fixes setting the subgroup size.
It also changes the behavior for workgroup sizes <=32, which now always
run with waveSize=32 and may have subgroupSize=64 (previously, the wave
size would have been forced to 64, with the upper 32 lanes inactive).

For testing, llpc and lgc tests failed with the new assert
`waveSize <= subgroupSize`.