Complete the dialect-ification of buffer pointers
Now that the transition to opaque pointers is complete, we can cleanly
represent the lgc.buffer.desc.to.ptr operation and tighten up the type
constraints on the other ops.

Also start using the llvm_dialects::Visitor as a type switch over
operations. The idea here is that in the long run, the type switch can
be optimized using something like a map lookup (this isn't implemented
yet). There is still room for improvement in this part: ideally, we'd
treat all instructions (native LLVM IR *and* dialect operations) as part
of the same visitor, but that needs more changes in llvm-dialects.