lgc: Add ModuleBunch infrastructure
This adds a pseudo-IR unit "ModuleBunch" (a bunch of modules), and the
pass manager infrastructure to allow module bunch passes and analysis.

In the same way as a module pass manager can have an adaptor to run
(e.g.) function passes, a module bunch pass manager can have an adaptor
to run module passes.

Currently that adaptor runs multiple modules sequentially, but the
design allows for a future change where modules can be run in parallel,
as long as they are created in separate LLVMContexts. If that happens, a
ModuleBunch pass that wants to move code from one module to another
would need to be aware of that possibility.

To run in parallel, there needs to be a separate copy of the
ModulePassManager inside the adaptor per thread. To enable that, there
is a constructor that takes a PassMaker function instead of a pass, and
the adaptor calls that once per distinct LLVMContext. There is no actual
parallelism yet.

The normal PassBuilder does not know about module bunch passes, so you
must instead use MbPassBuilder (added here).

Unlike for other real IR units, there is no way to directly get the
parent ModuleBunch from a Module. A user of this feature could provide
its own local way to do that for its passes, perhaps via a thread-local
variable.

LLVM's standard instrumentations unfortunately blow up in some places
when they see a novel IR unit they haven't heard of. This commit adds a
new MbStandardInstrumentations that contains copies of a few of the
standard instrumentations fixed up to cope with ModuleBunches. In
particular, -print-after etc and -verify-ir work.