builder: Perform wave64 bpermute in wwm
`ds_bperm` does not return a value if the source lane is inactive.
In wave64, this causes issues since we only use half of the exec mask.

For example:
Lane 37 (upper half) reads from lane 10. The exec mask
is `0x1FF FFFFFFFF`. Both lanes are active.
However, lane 37 uses `0x1FF` to determine if lane 10 is active.
Using this mask, it is not. Therefore the return value is 0.

Using Whole Wave Mode fixes the issue by setting
the permuted value of inactive lanes to 0 and
activating all lanes.