lgc: Add a TypeLowering utility
This is a helper class for lowering passes that want to replace values
of one type with values from another type (or tuples of values of other
types). It provides generic handling of phi nodes and a bunch of other
things, like taking care of cases where the lowered types are contained
in aggregates (structs or arrays).

The intention is to first use this for PatchBufferOp and then later
extend to other cases where "opaque" types are used.