Fix assert/trap in GS emit counter code (#2347)
I only saw a problem on Mac, not Linux, and then not on the first time
through this code. I think the problem was that the std::make_pair did a
copy of the m_emitCounterPtrs vector into a local temp, then the return
cast that by taking an ArrayRef of its contents. That would then become
invalid as soon as we returned from the function. I'm surprised it
worked as well as it did.