PatchBufferOp: fully convert to using llvm_dialects::Visitor and TypeLowering
This allows us to potentially merge this pass with other lowering passes
in the future. The idea is that each set of related dialect operations
that we want to lower can be "plugged into" a single visitor. A single pass
over the entire IR module can handle all of them, but we keep them
logically separation in the compiler source.

Generic instructions like phis are now handled by TypeLowering. A
complication is that we need to track the divergence of buffer
descriptors, and whether they refer to invariant buffers. This tracking
is now handled a bit more systematically. It would be even nicer if LLVM's
divergence analysis were updatable and handled this automatically.

v2:
- rebase
- fix a trivial oversight

v3:
- use (a wrapper around) std::optional<bool>
- fix a doc comment

v4:
- fix an issue with memcpy between different address spaces
- expand the common on visitPhiInst

v5:
- handle the case where a select instruction has a divergent condition

v6:
- fix an issue when propagating only one of invariant or divergent
- update with new tests