RegisterMetadataBuilder: optimize the metadata of Tgid{X,Y,Z}En for compute shader
Currently, Tgid{X,Y,Z}En are hard-coded as true unconditionally. There
is room for improvement if compute shader never reads some components of
gl_WorkgroupId, then that part could be set false. The part is
identified by the existence of ExtractElementInst as the user of
WorkgroupId input, and the result is recorded in term of function
attribute "amdgcn-no-workgroup-id-{x,y,z}".