PatchBufferOp: remove a nonsensical `load` case
We were handling loads of fat buffer pointers from the constant address
space, but in fact the load was only loading a buffer descriptor, which
conceptual doesn't make sense.

If a frontend wants to load a buffer descriptor as a fat buffer pointer,
it should use an explicit `lgc.buffer.desc.to.ptr` op.