Fix nullptr fat pointer handling in PatchBufferOp (#2363)
* Change the type of PatchBufferOp's m_replacementMap

Keys in this map are always Instructions so change the map type to
reflect that. No functional change intended.

* Add m_descType in PatchBufferOp

This just simplifies a few things. No functional change intended.

* Fix nullptr fat pointer handling in PatchBufferOp

PatchBufferOp had very limited support for nullptr fat pointer values in
getRemappedValueOrNull which only worked for values that were statically
known to be nullptr. There were special cases for handling these values
in visitSelectInst and replaceICmp.

Once nullptr values can leak into SelectInsts and PHINodes this is no
longer sufficient because we need to handle values that are not
statically known to be nullptr but may dynamically be nullptr. Fix this
by generating a real buffer descriptor value (I chose
<4 x i32> zeroinitializer) as well as a real index value for nullptrs.

This fixes a compiler crash in:
dEQP-VK.compute.workgroup_memory_explicit_layout.copy_memory.variable_pointers
where a recent change to the way we insert buffer descriptor loads was
enough to create a nullptr as an operand to a phi node, which is not one
of the "special cases" mentioned above and was not prepared to handle
it.