Remove state from Builder and BuilderRecorder
This takes advantage of the fact that we no longer support builder
direct mode; a Builder in the front-end is always a BuilderRecorder.

Most of the work here is to change BuilderRecorder so it does not hold a
PipelineState object or a ShaderModes object. Now, when the front-end
wants to write shader modes, it calls a static method on Pipeline, and
that calls on to a static method in ShaderModes to write the values
directly into IR, instead of caching in a ShaderModes object.

The middle-end still creates a ShaderModes object and reads the
applicable IR metadata into it on creation.

The point of this is that having no state in Builder or BuilderRecorder
(other than what is in IRBuilder) means that the front-end could just
construct a local Builder object when needed, instead of having to
construct a single Builder and get it into front-end passes.

Next steps in dismantling the BuilderRecorder/BuilderImpl thing would
be:
- remove 'virtual' from all the methods;
- separate the class hierarchy so Builder is BuilderRecorder, and
  BuilderImpl is a separate thing that does not inherit from Builder.