lgc: Builder/BuilderRecorder/BuilderImpl separation
Further leveraging not having builder direct mode any more:

* The BuilderImpl classes no longer have Builder in their class
  hierarchy, and no longer implement Builder's virtual methods. Instead,
  they have the same-named methods but unrelated in the class hierarchy.

* The BuilderImpl classes are now joined into one big BuilderImpl class,
  so each separate source file (ArithBuilder.cpp etc) can use methods in
  the other one without using the common virtual functions in Builder.

* Builder no longer needs to have virtual methods. Builder always
  records, and all the methods in BuilderRecorder are moved into
  Builder.

* BuilderRecorder still exists as a class, but only to hold a couple of
  static methods for opcode <-> name conversion that are also used in
  BuilderReplayer.

Not included in this commit, the front-end can now change to create a
Builder as a local variable or its own "new", instead of having to use
LgcContext::CreateBuilder. It can create one wherever it needs it,
instead of having to create a single one for the whole front-end and get
it into passes.