Optimize unspecified channel to 0.0 when exporting vertex attributes (#2359)
* Initialize outputs to zero if the client requires

Add to `Vkgc::PipelineShaderOptions` a field
`workgroundInitializeOutputsToZero` to tell the compiler whether we need
initialize outputs to zero or not.
Adjust `transValueWithOpcode<OpVariable>` to create an `zeroinitializer`
if the user-defined output global doesn’t have an initializer and this
option is set.

* Reduce useless export by using DEFAULT_VAL for the attribute.

Add to `FsInterpInfo` a field `isDefaultVal`.
The optimization is done in `PatchInOutImportExport::exportVertexAttribs`
by checking whether all the export components are constant as either
undef or 0.0. If they are, skip the export and instead set `
resUsage->inOutUsage.fs.interpInfo[i].isDefaultVal`.