eliminate compile warning
 + tool/vfx/vfxParser.cpp:1368:15: warning: `sprintf` argument 4 may overlap destination object `namePos` [-Wrestrict]

Currently this function never run it, because macroDefinition is always empty.
In fact, lineRest may OVERLAP with the first parameter (namePos).

The C standard and POSIX specify that the behavior of sprintf and its variants
is undefined when an argument overlaps with the destination buffer.

Reference: https://en.cppreference.com/w/c/io/fprintf