eliminate compile warning
 + tool/vfx/vfxVkSection.h:497:52: warning: `offsetof` within non-standard-layout type `Vfx::SectionIndirectCalleeSavedRegs` is conditionally-supported [-Winvalid-offsetof]

If type is not a standard layout type, the result of offsetof is undefined.
Reason:
 + has virtual base classes,
 + has non-standard-layout base classes, and
 + multiple classes in the hierarchy has non-static data members.

Reference: https://en.cppreference.com/w/cpp/language/classes#Standard-layout_class

We often use the following code in C language to calculate the offset of a
member. That is, address of member is offset!
```c
((size_t) &((TYPE *)0)->MEMBER)
```