Handle more complex waterfalls when scalarizing descriptor loads.
1. Guard uses of comesBefore when scalarizing descriptor loads.
Instruction::comesBefore is only valid on instructions in the same block.
However its possible we might accidentally call it inter-block when
determining if we can scalarize descriptor loads for a waterfall loop.
Inter-block waterfalls are not supported by backend.

2. Merge overlapping waterfalls with common descriptors.
If multiple operations share the same descriptors then we need to merge
the waterfalls, rather than creating unnecessary nesting.