Run inliner pass before lower ray tracing pass for all RT shader stage
Any ray tracing shader may access shader record buffer in non-entry function,
we will replace all Global SRB to a combination of instructions at the
beginning of entry function in SpirvLowerRayTracing pass, so we need to inline
all functions so that SRB is accessible throughout the shader.

e.g.

```
@ShaderRecordBuffer = ...

def entry() {
  access @ShaderRecordBuffer
  call foo()
}

def foo() {
  access @ShaderRecordBuffer
}
```

If we don't inline `foo()`, when we replaceAllUse of `@ShaderRecordBuffer`,
the one in `foo()` will be invalid.