Add 'component' argument to loadValueFromGsVsRing
The reason is that we ought to be able to load component value from a
specific output location. Currently, we always load the whole value of a
specific output location. In the future, we want to just load the
necessary component value.

Will have more work on PatchCopyShader::exportOutput
in follow-up changes. This is just a preparation change.